refactor(Command): command를 컨트롤러로 분류
- 클래스를 컨트롤러로 분류
- controller의 if문 제거 후, command에서 명령어에 따라 수행하도록 수정
