+++ src/main/java/chess/controller/ChessController.java
@@ -1,6 +1,5 @@
 package chess.controller;
 
-import chess.domain.game.Command;
 import chess.domain.game.GameState;
 import chess.domain.game.Ready;
 import chess.dto.BoardAndTurnInfo;
@@ -8,6 +7,8 @@ import chess.dto.Request;
 import chess.dto.ScoreResponse;
 import chess.view.InputView;
 import chess.view.OutputView;
+
+import java.util.List;
 
 public class ChessController {
 
@@ -28,32 +29,32 @@ public class ChessController {
         try {
             Request request = Request.of(InputView.inputCommand());
             Command command = Command.find(request.getCommand());
-            return runGameByCommand(gameState, request, command);
+            return command.execute(gameState, request.getArguments());
         } catch (RuntimeException e) {
             OutputView.printException(e);
             return tryExecute(gameState);
         }
     }
 
-    private GameState runGameByCommand(GameState gameState, Request request, Command command) {
-        if (command.isType(Command.START)) {
-            GameState state = gameState.start();
-            OutputView.printBoardAndTurn((BoardAndTurnInfo) state.getResponse());
-            return state;
-        }
-        if (command.isType(Command.FINISH)) {
-            return gameState.finish();
-        }
-        if (command.isType(Command.MOVE)) {
-            GameState state = gameState.move(request.getArguments());
-            OutputView.printBoardAndTurn((BoardAndTurnInfo) state.getResponse());
-            return state;
-        }
-        if (command.isType(Command.STATUS)) {
-            GameState state = gameState.status();
-            OutputView.printStatus((ScoreResponse) state.getResponse());
-            return state;
-        }
-        return null;
+    static GameState start(GameState gameState, List<String> ignored) {
+        GameState state = gameState.start();
+        OutputView.printBoardAndTurn((BoardAndTurnInfo) state.getResponse());
+        return state;
+    }
+
+    static GameState finish(GameState gameState, List<String> ignored) {
+        return gameState.finish();
+    }
+
+    static GameState move(GameState gameState, List<String> arguments) {
+        GameState state = gameState.move(arguments);
+        OutputView.printBoardAndTurn((BoardAndTurnInfo) state.getResponse());
+        return state;
+    }
+
+    static GameState status(GameState gameState, List<String> ignored) {
+        GameState state = gameState.status();
+        OutputView.printStatus((ScoreResponse) state.getResponse());
+        return state;
     }
 }
